backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyMuPDF==1.23.8
httpx==0.26.0
websockets==12.0
google-generativeai==0.3.2
pageindex>=0.1.0
pydantic==2.5.3
